+++ projects/ontimize-web-ngx-report/src/lib/components/index.ts
@@ -1,0 +1,2 @@
+export * from './report/index';
+export * from './report-on-demand/index';
