+++ projects/ontimize-web-ngx-report/src/lib/components/report-on-demand/index.ts
@@ -1,0 +1,5 @@
+export * from './apply-configuration/apply-configuration-dialog.component';
+export * from './report-on-demand/report-on-demand.component';
+export * from './save-preferences-dialog/save-preferences-dialog.component';
+export * from './select-function-dialog/select-function-dialog.component';
+export * from './style-dialog/style-dialog.component'
