+++ projects/ontimize-web-ngx-report/src/lib/components/report/index.ts
@@ -1,0 +1,4 @@
+export * from './o-report-detail/o-report-detail.component';
+export * from './o-report-home/o-report-home.component';
+export * from './o-report-new/o-report-new.component';
+export * from './o-report-viewer/o-report-viewer.component';
