+++ projects/ontimize-web-ngx-report/src/lib/components/report/o-report-detail/o-report-detail.component.ts
@@ -1,10 +1,9 @@
 import { Component, ViewChild } from '@angular/core';
 import { MatDialog } from '@angular/material';
 import { DialogService, OFormComponent, OTextInputComponent } from 'ontimize-web-ngx';
+import { DEFAULT_HEIGHT_DIALOG } from '../../report-on-demand';
 import { OReportViewerComponent } from '../o-report-viewer/o-report-viewer.component';
 
-export const DEFAULT_WIDTH_DIALOG = '70%';
-export const DEFAULT_HEIGHT_DIALOG = '90%';
 @Component({
   selector: 'o-report-detail',
   templateUrl: './o-report-detail.component.html'
