+++ projects/ontimize-web-ngx-report/src/lib/components/report/o-report-viewer/o-report-viewer.component.ts
@@ -1,12 +1,11 @@
 import { Component, Inject, OnInit } from '@angular/core';
 import { MAT_DIALOG_DATA } from '@angular/material';
 import { DialogService } from 'ontimize-web-ngx';
-import { OReportService } from '../o-report.service';
+import { OReportService } from '../../../services/o-report.service';
 
 @Component({
   selector: 'o-report-viewer',
-  templateUrl: './o-report-viewer.component.html',
-  styleUrls: ['./o-report-viewer.component.scss']
+  templateUrl: './o-report-viewer.component.html'
 })
 export class OReportViewerComponent implements OnInit {
 
@@ -16,9 +15,9 @@ export class OReportViewerComponent implements OnInit {
     @Inject('report') private reportService: OReportService,
     protected dialogService: DialogService,
     @Inject(MAT_DIALOG_DATA) public data: any) {
-      this.reportService.configureService(this.reportService.getDefaultServiceConfiguration());
-      this.reportService.configureAdapter();
-      this.reportService.fillReport(this.data['params'], 'fillReport', {}, this.data['filter']).subscribe(
+    this.reportService.configureService(this.reportService.getDefaultServiceConfiguration());
+    this.reportService.configureAdapter();
+    this.reportService.fillReport(this.data['params'], 'fillReport', {}, this.data['filter']).subscribe(
       res => {
         if (res && res.data.length && res.code === 0) {
           this.pdf = res.data[0].file;
@@ -28,13 +27,14 @@ export class OReportViewerComponent implements OnInit {
         if (this.dialogService) {
           this.dialogService.error('ERROR',
             'SERVER_ERROR_MESSAGE');
-          }
-          console.log(err);
+        }
+        console.log(err);
       }
     );
-    }
+  }
 
   ngOnInit() {
+    //ngOnInit implementation
   }
