+++ projects/ontimize-web-ngx-report/src/lib/i18n/i18n.ts
@@ -1,4 +1,4 @@
-export var MAP = {
+export let MAP = {
 
   // English
   en: {
