+++ projects/ontimize-web-ngx-report/src/lib/services/index.ts
@@ -1,0 +1,5 @@
+export * from './o-alert.service';
+export * from './o-fill-report.service';
+export * from './o-report.service';
+export * from './reports-on-demand.service';
+export * from './reports.service';
