+++ projects/ontimize-web-ngx-report/src/lib/services/o-fill-report.service.ts
@@ -1,7 +1,7 @@
 import { Inject, Injectable } from '@angular/core';
 import { MatDialog } from '@angular/material';
-import { DialogService, Util } from 'ontimize-web-ngx';
-import { OReportViewerComponent } from './o-report-viewer/o-report-viewer.component';
+import { DialogService } from 'ontimize-web-ngx';
+import { OReportViewerComponent } from '../components/report/o-report-viewer/o-report-viewer.component';
 import { OReportService } from './o-report.service';
 
 @Injectable()
@@ -16,7 +16,7 @@ export class OFillReportService {
   openFillReport(reportId: string, parametersValues: object, filter: object) {
     this.reportService.configureService(this.reportService.getDefaultServiceConfiguration());
     this.reportService.configureAdapter();
-    let kv = {'UUID': reportId};
+    let kv = { 'UUID': reportId };
     this.reportService.query(kv, null, 'getReport', {}).subscribe(
       res => {
         if (res && res.data.length && res.code === 0) {
@@ -24,7 +24,7 @@ export class OFillReportService {
           if (parameters.length > 0) {
             let av = [reportId];
             let values = Object.values(parametersValues);
-            for (let i=0; i<values.length; i++) {
+            for (let i = 0; i < values.length; i++) {
               av.push(values[i]);
             }
             this.dialog.open(OReportViewerComponent, {
@@ -51,9 +51,9 @@ export class OFillReportService {
       err => {
         if (this.dialogService) {
           this.dialogService.error('ERROR',
-              'SERVER_ERROR_MESSAGE');
-          }
-          console.log(err);
+            'SERVER_ERROR_MESSAGE');
+        }
+        console.log(err);
       }
     );
