+++ projects/ontimize-web-ngx-report/src/lib/types/index.ts
@@ -1,0 +1,5 @@
+export * from './report-column-style.type';
+export * from './report-configuration.type';
+export * from './report-function.type';
+export * from './report-orderBy.type';
+export * from './report-preferences.type';
