+++ projects/ontimize-web-ngx-report/src/lib/util/index.ts
@@ -1,0 +1,1 @@
+export * from './o-reports-translate.pipe';
