+++ projects/ontimize-web-ngx-report/src/public-api.ts
@@ -1,7 +1,8 @@
 /*
  * Public API Surface of ontimize-web-ngx-report
  */
-
 export * from './lib/ontimize-web-ngx-report.module';
-export * from './lib/components/report/o-report.service';
-export * from './lib/components/report/o-fill-report.service';
+export * from './lib/components/index';
+export * from './lib/services/index';
+export * from './lib/types/index';
+export * from './lib/util/index';
